`php-get-current-class` matches valid class name
`php-get-current-class` had been dropping any ".php" suffix from the file/buffer name and assuming that the remainder was the name of the class.  This was not the best method of seeking the class name.  As the syntax of valid PHP class names is known[1], a more flexible method was to seek such a class name in the filename.  `php-get-current-class` now returns the first matching regexp in the filename as the name of the class.

[1] http://www.php.net/manual/en/language.variables.basics.php