(phpunit-get-current-class) returns exact unit test classname
The new routine preferably performs a backwards regex to find the start
of the current class;  if that fails, or the 'file' parameter has been
set, it falls back to the filename given or the filename of the current
buffer.  It ensures that the derived classname ends in the word 'Test',
which I believe to be a required PHPUnit convention.

The regular expression passed to --filter also matches any namespace
that might be found;  this may be unnecesary, I'm still looking into
that.  It can be disregarded for now.