(phpunit-current-class) finds exact unit test
This commit is a candidate for merging with the master branch, so a
summary is in order.  Two convenience regexps have been defined as
constants: `php-labelchar-regexp` matches the characters in a valid php
label, and `php-beginning-of-class` seeks the name at the beginning of a
class.  These are used by the rewritten `phpunit-get-current-class` and
`php-current-class` functions.

The `phpunit-get-current-class` function now attempts to identify the
current class from a backwards regexp, but falls back on the buffer name
if that fails.  In either case, it ensures that the final part of the
resulting classname is 'Test', requiring PHPUnit's convention that test
unit classes be named *Class*Test, where 'Class' is the name of the
class under test.

The `phpunit-current-class` function now passes a lookbehind assertion
in the '--filter' option to phpunit.  The purpose of this is to ensure
that no valid php label character can precede the rest of the string
passed to PHPUnit when it is matching against the filter expression.
This, along with enforcing the presence of 'Test' at the end of the
string, should force the isolation of the exact correct unit tests for
any given class.

Unit tests for the revised features of `phpunit-get-current-class` have
been written and included.