feat: Rerun last test even from different phpunit root dir
This comes up, for example, if the code being worked on is in a package but the
test is being run within a dependent project, where they each have their own
phpunit.xml. This could be where a package is being developed locally and is
symlinked into an app's vendor directory, or in a monorepo setting where the
package and project may be interdependent.

Signed-off-by: Clayton Carter <claytonrcarter@gmail.com>